backend/src/requirements.txt
# file path: backend/src/requirements.txt
yfinance
pandas
numpy
google-genai